+++ data/state-feed.json
@@ -3,12 +3,12 @@
   "feed_id": "paper-daily",
   "last_run_date": "2026-05-31",
   "last_preferred_date": "2026-05-31",
-  "last_run_at": "2026-06-02T08:13:07.466549+00:00",
+  "last_run_at": "2026-06-02T10:42:53.712062+00:00",
   "last_attempted_dates": [
     "2026-05-31"
   ],
   "last_update_status": "updated",
-  "last_success_at": "2026-06-02T08:13:07.466549+00:00",
+  "last_success_at": "2026-06-02T10:42:53.712062+00:00",
   "last_success_date": "2026-05-31",
   "last_content_date": "2026-05-31",
   "latest_updates_date": "2026-05-31",
