+++ feed-papers.json
@@ -1,7 +1,7 @@
 {
   "schema_version": "v1",
   "feed_id": "paper-daily",
-  "generated_at": "2026-06-02T08:13:07.455453+00:00",
+  "generated_at": "2026-06-02T10:42:53.708065+00:00",
   "run_date": "2026-05-31",
   "source_repo": "xianshang33/llm-paper-daily",
   "items": [
